cpu: Migrate from rv64 to rv32

+++ app/crates/kartoffels-cpu-tests/src/lib.rs
@@ -1,6 +1,6 @@
-#![cfg_attr(target_arch = "riscv64", no_std)]
+#![cfg_attr(target_arch = "riscv32", no_std)]
 
-#[cfg(target_arch = "riscv64")]
+#[cfg(target_arch = "riscv32")]
 #[macro_export]
 macro_rules! test {
     ($code:literal) => {
@@ -13,7 +13,7 @@ macro_rules! test {
     };
 }
 
-#[cfg(not(target_arch = "riscv64"))]
+#[cfg(not(target_arch = "riscv32"))]
 #[macro_export]
 macro_rules! test {
     ($code:literal) => {
@@ -23,17 +23,17 @@ macro_rules! test {
     };
 }
 
-#[cfg(target_arch = "riscv64")]
+#[cfg(target_arch = "riscv32")]
 pub fn exit(val: u32) {
     use core::arch::asm;
 
     unsafe {
-        asm!("mv x10, {}", in(reg) val as u64);
+        asm!("mv x10, {}", in(reg) val);
         asm!("ebreak");
     }
 }
 
-#[cfg(not(target_arch = "riscv64"))]
+#[cfg(not(target_arch = "riscv32"))]
 pub fn exit(val: u32) {
     println!("{val}");
 }
